Move detect system to its own module
Moving more content from the detect script to specific module to
separate parsing logic.
